Adding navigation type and toggling based on screen state.
