+++ app/src/main/java/com/adammcneilly/reader/ui/ReaderNavigationType.kt
@@ -1,0 +1,6 @@
+package com.adammcneilly.reader.ui
+
+enum class ReaderNavigationType {
+    BOTTOM_NAVIGATION,
+    NAVIGATION_RAIL,
+}
